# openapi/egdata.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "egdata.app API",
    "version": "0.1.0-alpha.0",
    "description": "REST API for Epic Games Store catalog, pricing, search, and public egdata.app data. This contract starts with public stable REST endpoints and will expand as legacy routes are classified.",
    "contact": {
      "name": "egdata.app",
      "url": "https://egdata.app"
    }
  },
  "servers": [
    {
      "url": "https://api.egdata.app",
      "description": "Production"
    },
    {
      "url": "https://api-gcp.egdata.app",
      "description": "Production GCP mirror"
    },
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/offers/{id}/items": {
      "get": {
        "operationId": "listOfferItems",
        "tags": [
          "Offer Details"
        ],
        "summary": "List items for an offer",
        "parameters": [
          {
            "$ref": "#/components/parameters/offerId"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/items": {
      "get": {
        "operationId": "listItems",
        "tags": [
          "Items"
        ],
        "summary": "List catalog items",
        "description": "Returns paginated Epic catalog items sorted by last modification date.",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "operationId": "getItem",
        "tags": [
          "Items"
        ],
        "summary": "Get an item by ID",
        "description": "Looks up an item by Mongo document ID or Epic item ID and expands custom attributes.",
        "parameters": [
          {
            "$ref": "#/components/parameters/itemId"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/items/bulk": {
      "post": {
        "operationId": "getItemsBulk",
        "tags": [
          "Items"
        ],
        "summary": "Fetch multiple items by ID",
        "description": "Fetches up to 100 item IDs in one request.",
        "requestBody": {
          "required": true,
          "description": "Item IDs to fetch.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "items"
                ],
                "additionalProperties": false,
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/items/bulk/offers": {
      "post": {
        "operationId": "getOffersForItemsBulk",
        "tags": [
          "Items"
        ],
        "summary": "Resolve the best offer for multiple items",
        "description": "Returns an object keyed by requested item ID. Each value is the best matching offer or null when no offer can be resolved.",
        "parameters": [
          {
            "$ref": "#/components/parameters/locale"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Item IDs to resolve.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "items"
                ],
                "additionalProperties": false,
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Offer"
                      }
                    ],
                    "nullable": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/items/{id}/assets": {
      "get": {
        "operationId": "getItemAssets",
        "tags": [
          "Items"
        ],
        "summary": "List assets for an item",
        "parameters": [
          {
            "$ref": "#/components/parameters/itemId"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/items/{id}/builds": {
      "get": {
        "operationId": "getItemBuilds",
        "tags": [
          "Items"
        ],
        "summary": "List builds associated with an item",
        "parameters": [
          {
            "$ref": "#/components/parameters/itemId"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Build"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/items/{id}/changelog": {
      "get": {
        "operationId": "getItemChangelog",
        "tags": [
          "Items"
        ],
        "summary": "List changelog entries for an item",
        "parameters": [
          {
            "$ref": "#/components/parameters/itemId"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChangelogEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/items/{id}/offer": {
      "get": {
        "operationId": "getItemOffer",
        "tags": [
          "Items"
        ],
        "summary": "Get the offer associated with an item",
        "parameters": [
          {
            "$ref": "#/components/parameters/itemId"
          },
          {
            "$ref": "#/components/parameters/locale"
          }
        ],
        "x-egdata-visibility": "public",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Offer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "EpicSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "egdata.app browser session backed by Epic authentication."
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Endpoint-specific API key."
      }
    },
    "parameters": {
...
      "locale": {
        "name": "locale",
        "in": "query",
        "required": false,
        "description": "Exact BCP-47-style locale used to overlay localized public offer text fields. Defaults to en-US.",
        "schema": {
          "type": "string",
          "example": "es-ES"
        }
      },
      "limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of records to return. Most list endpoints cap this server-side.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "example": 10
        }
      },
      "page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "One-based page number.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "offerId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Epic offer identifier.",
        "schema": {
          "type": "string",
          "example": "09176f4ff7564bbbb499bbe20bd6348f"
        }
      },
      "itemId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Epic item identifier.",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "Item": {
        "type": "object",
        "description": "Public Epic Games Store item DTO.",
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "string"
          },
          "namespace": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "keyImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "releaseInfo": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "ItemListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "elements",
          "page",
          "limit",
          "total"
        ],
        "properties": {
          "elements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "Image": {
        "type": "object",
        "description": "Epic image asset reference.",
        "additionalProperties": true,
        "properties": {
          "type": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}